added Docker, License and dumped sql for MySQL.
Recently I was trying to configure Dockerfile and docker-compose for the
first time. Eventually, I've got to the point where I can start my
application with just one command: docker-compose up.
Also, I saw license in one of WM repositories and decided that it would
be cool to add some here.. I know, I know.
And I as well completed README for starting app with Docker or without
it.